piecrust: refactored and improved performance
